requirements.txt
asgiref             3.9.1
blinker             1.9.0
click               8.2.1
colorama            0.4.6
Django              5.2.5
djangorestframework 3.16.1
itsdangerous        2.2.0
MarkupSafe          3.0.2
pip                 25.2
psycopg2-binary     2.9.10
PyJWT               2.10.1
python-dotenv       1.1.1
sqlparse            0.5.3
tzdata              2025.2
Werkzeug            3.1.3